Added metadata that allows modders to guarantee custom version of SMM
Some harmless metadata gets added to the (text_)misc.xml file when
at least one mod is patched. This metadata will then show up on
every patch, allowing mods to require it, so that the previous versions
of SMM will not be able to patch it, while this one and later will.

Example of this and possibly ability to turn this feature off forthcoming.