Added support for XSLT stylesheets
Mods can incorporate .xsl files to transform an existing file
in the currently patched dat.

For example, events.xsl will target the existing events.xml,
but if there's no blah.xml, blah.xsl won't do anything.

User discretion is currently required to know if the transform
works or breaks the file. And so this is just an initial commit
to prove functionality; further revisions will be forthcoming.