Better logging and better sandbox search; up version to 1.10.1
When patching transforms you now have more information to see how
exactly the patch process is going. It might be a bit too much
info for the end user, but for now it's good details.

Searching files opened and patched using XML Sandbox now has option
of case-sensitivity as well as using a regular expression, both should
result in faster testing; the latter of which should also make
testing easier on find tags with regex="true".