Add an option for rebuildXMLFile to add root FTL tags
While it's not used by Slipstream itself, this is useful for
Project Wormhole, which uses large parts of Slipstream code for it's own
mod loading.

Upstreaming this patch helps Wormhole keep it's copy of Slipstream
up-to-date, as it removes a patch that would otherwise need to be
applied whenever Slipstream is updated.