Added support for Dino's Mac Hyperspace wrapper
Slipstream works by knowing the location of the dat,
patching it, and running the game. On Windows, the location is
easy to determine because the exe and dat are in the same folder.

However, on Mac, the app itself runs the game, and the dat is
within the app. So when Slipstream is given the location of the
app, it knows the folder structure to succesfully locate the dat.

The wrapper .app has a different internal folder structure, so
new code was needed to properly locate the dat. Manually finding
the app has been tested to work; hopefully automatic works too.