Fixed running Windows .exe with later Java versions; minimum Java now 8
Users with only newer versions of Java were unable to run the .exe on Windows.
Newer versions appear not to add registry keys by default. For this reason,
Launch4j 3.50+ defaults to viewing PATH and JAVA_HOME, hence, new .exe's.

The minimum Java version is now 8 because it is unlikely there are any
users running 6 or 7. Project will likely be updated in general to Java 8.
---------
Co-authored-by: blizzarchon <gopengy@gmail.com>